pydub-core/pyproject.toml
[project]
name = "pydub-core"
version = "0.1.0"
requires-python = ">=3.11"

[dependency-groups]
dev = ["pytest-benchmark>=5.1.0"]

[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[tool.maturin]
python-source = "python"
module-name = "pydub_core._pydub_core"
